fix: keep route page exports component-only
