+++ apps/web/src/routes/_authenticated/appointments/-appointment-detail-data.ts
@@ -1,0 +1,2 @@
+export const APPOINTMENT_DETAIL_RESOURCE_PAGE_SIZE = 25
+export const AVAILABLE_HAIR_ORDERS_PAGE_SIZE = 100
