+++ apps/web/src/routes/_authenticated/customers/index.tsx
@@ -1,6 +1,7 @@
 import { createFileRoute } from "@tanstack/react-router"
 
-import { customersListQueryOptions, CustomersPage, searchSchema } from "./-index-page"
+import { customersListQueryOptions, searchSchema } from "./-index-data"
+import { CustomersPage } from "./-index-page"
 
 export const Route = createFileRoute("/_authenticated/customers/")({
   component: CustomersPage,
