+++ apps/web/src/routes/_authenticated/documents/$documentId/match.tsx
@@ -1,12 +1,12 @@
 import { createFileRoute, redirect } from "@tanstack/react-router"
 
 import {
-  DocumentMatchPage,
   documentQueryOptions,
   MATCH_CANDIDATES_PAGE_SIZE,
   matchCandidatesQueryOptions,
   searchSchema,
-} from "./-match-page"
+} from "./-match-data"
+import { DocumentMatchPage } from "./-match-page"
 
 export const Route = createFileRoute("/_authenticated/documents/$documentId/match")({
   component: DocumentMatchPage,
