+++ apps/web/src/routes/_authenticated/documents/-index-page.tsx
@@ -19,7 +19,6 @@ import { IconLinkOff, IconTrash } from "@tabler/icons-react"
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 import { Link } from "@tanstack/react-router"
 import { useState } from "react"
-import { z } from "zod"
 
 import { AttachmentPreviewDialog, type AttachmentPreview } from "@/components/attachment-preview-dialog"
 import { BreadcrumbItem } from "@/components/breadcrumbs"
@@ -28,23 +27,8 @@ import { Section } from "@/components/section"
 import { type Currency, formatMinor } from "@/lib/currency"
 import { trpc } from "@/utils/trpc"
 
+import { type DocumentStatus, documentsQueryOptions, PAGE_SIZE } from "./-index-data"
 import { Route } from "./index"
-
-export const PAGE_SIZE = 25
-const documentStatusSchema = z.enum(["unassigned", "assigned", "all"])
-export const searchSchema = z.object({
-  page: z.coerce.number().int().min(1).optional(),
-  status: documentStatusSchema.optional(),
-})
-
-type DocumentStatus = z.infer<typeof documentStatusSchema>
-
-export function documentsQueryOptions(input: { status: DocumentStatus; page: number }) {
-  return trpc.bankStatementAttachments.list.queryOptions(
-    { assignmentStatus: input.status, page: input.page, pageSize: PAGE_SIZE },
-    { placeholderData: (previousData) => previousData },
-  )
-}
 
 export function DocumentsPage() {
   const search = Route.useSearch()
