+++ apps/web/src/routes/_authenticated/hair-orders/-$hairOrderId-data.ts
@@ -1,0 +1,1 @@
+export const AVAILABLE_HAIR_ORDERS_PAGE_SIZE = 100
